Replace conversion operators from Value/GridPos with conversion methods

+++ src/actors/KillerActor.cc
@@ -48,7 +48,7 @@ namespace enigma {
     }
     
     int Killer::traitsIdx() const {
-        return getAttr("color");
+        return getAttr("color").toInt();
     }
 
     ActorTraits Killer::traits[2] = {
